refactor(cli): extraer capa de servicios para run/build/test/mod
