+++ src/pcobra/cobra/cli/commands_v2/run_cmd.py
@@ -6,7 +6,7 @@ from pcobra.cobra.architecture.backend_policy import PUBLIC_BACKENDS
 from pcobra.cobra.architecture.contracts import assert_backend_allowed_for_scope
 from pcobra.cobra.cli.commands.base import BaseCommand
 from pcobra.cobra.bindings.runtime_manager import RuntimeManager
-from pcobra.cobra.cli.services.command_factory import CommandFactory
+from pcobra.cobra.cli.services.run_service import RunService
 from pcobra.cobra.cli.i18n import _
 from pcobra.cobra.cli.utils.messages import mostrar_error
 from pcobra.cobra.cli.utils.autocomplete import files_completer
@@ -20,8 +20,7 @@ class RunCommandV2(BaseCommand):
 
     def __init__(self) -> None:
         super().__init__()
-        self._command_factory = CommandFactory()
-        self._legacy = self._command_factory.create("ejecutar")
+        self._service = RunService()
         self._runtime_manager = RuntimeManager()
 
     def register_subparser(self, subparsers: Any):
@@ -66,4 +65,4 @@ class RunCommandV2(BaseCommand):
             modo=getattr(args, "modo", "mixto"),
             backend_reason=resolution.reason_for(debug=debug),
         )
-        return self._legacy.run(legacy_args)
+        return self._service.run(legacy_args)
