+++ src/pcobra/cobra/cli/commands_v2/test_cmd.py
@@ -5,7 +5,7 @@ from pcobra.cobra.build import backend_pipeline
 from pcobra.cobra.cli.commands.base import BaseCommand
 from pcobra.cobra.architecture.contracts import assert_backend_allowed_for_scope
 from pcobra.cobra.bindings.runtime_manager import RuntimeManager
-from pcobra.cobra.cli.services.command_factory import CommandFactory
+from pcobra.cobra.cli.services.test_service import TestService
 from pcobra.cobra.cli.i18n import _
 from pcobra.cobra.cli.utils.messages import mostrar_error
 from pcobra.cobra.cli.target_policies import (
@@ -23,8 +23,7 @@ class TestCommandV2(BaseCommand):
 
     def __init__(self) -> None:
         super().__init__()
-        self._command_factory = CommandFactory()
-        self._legacy = self._command_factory.create("verificar")
+        self._service = TestService()
         self._runtime_manager = RuntimeManager()
         self._default_langs = ",".join(VERIFICATION_EXECUTABLE_TARGETS)
 
@@ -75,4 +74,4 @@ class TestCommandV2(BaseCommand):
             modo=getattr(args, "modo", "mixto"),
             backend_reason=resolution.reason_for(debug=debug),
         )
-        return self._legacy.run(legacy_args)
+        return self._service.run(legacy_args)
