+++ src/pcobra/cobra/cli/services/__init__.py
@@ -1,5 +1,15 @@
 """Servicios reutilizables para comandos CLI."""
 
+from pcobra.cobra.cli.services.build_service import BuildService
 from pcobra.cobra.cli.services.command_factory import CommandFactory
+from pcobra.cobra.cli.services.mod_service import ModService
+from pcobra.cobra.cli.services.run_service import RunService
+from pcobra.cobra.cli.services.test_service import TestService
 
-__all__ = ["CommandFactory"]
+__all__ = [
+    "BuildService",
+    "CommandFactory",
+    "ModService",
+    "RunService",
+    "TestService",
+]
